requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pymongo==4.6.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pyzbar==0.1.9
pandas==2.1.4
python-dateutil==2.8.2
